Add tex2img package
This commit adds a new package called tex2img, which can be used to compile
LaTeX documents into PDF and PNG formats. The package includes a LatexCompiler
class, which can be used to compile LaTeX documents into PDF format, and a
Latex2PNG class, which can be used to compile LaTeX documents into PNG format.

In addition to the main package files, this commit also includes a setup.py
file for packaging and distribution, as well as some basic tests in the
tests directory.

This is an initial commit, and the package is still in development. The API
may change in the future, and more features and tests will be added in future
commits.