UChange LICENSE and README.md to use Apache-2.0 license
This commit updates the LICENSE and README.md files to use the Apache-2.0
license. The previous license was [yyyy], which has been replaced with the
Apache-2.0 license.

In addition, the README.md file has been created and filled out with a brief
introduction to the project, installation instructions, usage examples,
and information about contributing.

This commit does not modify any of the project code.